Move store and load to one file
